Resolve "Authentifizierung und Authorisierung muss einbauen"

Closes #167, #143, and #173

See merge request beba3606/vetlib!130

+++ backend/src/service/animalService.ts
@@ -3,6 +3,8 @@ import { animals, Prisma } from "../../generated/prisma";
 import { AnimalsCreateType, AnimalsType, AnimalUpdateType } from "vetlib-shared/schemas/ZodSchemas";
 import fs from "node:fs/promises";
 import path from "node:path";
+import { ResourceNotFoundError } from "../exceptions/errors/ResourceNotFoundError";
+import { ConstraintError } from "../exceptions/errors/ContraintError";
 
 export const animalService = {
   async create(data: AnimalsCreateType): Promise<AnimalsType> {
@@ -27,10 +29,19 @@ export const animalService = {
     });
 
     return ({
-      ...created,
+      dateofbirth: created.dateofbirth,
+      dateofbirthisexact: created.dateofbirthisexact,
+      heightincm: created.heightincm,
+      id: created.id,
+      iscastrated: created.iscastrated,
+      lifestyleisindoors: created.lifestyleisindoors,
+      name: created.name,
+      sex: created.sex,
+      timeofdeath: created.timeofdeath,
+      weightingram: created.weightingram,
       animalgroupid: created.fk_animalgroupid,
       animaltypeid: created.fk_animaltypeid
-    })
+    });
   },
 
   async update(data: AnimalUpdateType): Promise<AnimalsType> {
@@ -52,21 +63,45 @@ export const animalService = {
     });
 
     return ({
-      ...updated,
+      dateofbirth: updated.dateofbirth,
+      dateofbirthisexact: updated.dateofbirthisexact,
+      heightincm: updated.heightincm,
+      id: updated.id,
+      iscastrated: updated.iscastrated,
+      lifestyleisindoors: updated.lifestyleisindoors,
+      name: updated.name,
+      sex: updated.sex,
+      timeofdeath: updated.timeofdeath,
+      weightingram: updated.weightingram,
       animalgroupid: updated.fk_animalgroupid,
       animaltypeid: updated.fk_animaltypeid
-    })
+    });
   },
 
-  async getById(id: number): Promise<animals> {
+  async getById(id: number): Promise<AnimalsType> {
     const foundAnimal = await prisma.animals.findUnique({ where: { id } });
 
-    if (!foundAnimal) throw new Error(`Animal not found with id: ${id}`);
+    if (!foundAnimal) {
+      throw new ResourceNotFoundError(`Animal not found with id: ${id}`, 'id', id);
+    }
 
-    return foundAnimal;
+    return ({
+      dateofbirth: foundAnimal.dateofbirth,
+      dateofbirthisexact: foundAnimal.dateofbirthisexact,
+      heightincm: foundAnimal.heightincm,
+      id: foundAnimal.id,
+      iscastrated: foundAnimal.iscastrated,
+      lifestyleisindoors: foundAnimal.lifestyleisindoors,
+      name: foundAnimal.name,
+      sex: foundAnimal.sex,
+      timeofdeath: foundAnimal.timeofdeath,
+      weightingram: foundAnimal.weightingram,
+      animalgroupid: foundAnimal.fk_animalgroupid,
+      animaltypeid: foundAnimal.fk_animaltypeid
+    });
   },
 
-  async getByPersonId(personId: number): Promise<animals[]> {
+  async getByPersonId(personId: number): Promise<AnimalsType[]> {
     const relations = await prisma.person_has_animal.findMany({
       where: {
         fk_personid: personId
@@ -76,22 +111,83 @@ export const animalService = {
       }
     });
 
-    return relations.map(r => r.animals);
+    return relations.map(r => r.animals).flatMap(animal => ({
+      dateofbirth: animal.dateofbirth,
+      dateofbirthisexact: animal.dateofbirthisexact,
+      heightincm: animal.heightincm,
+      id: animal.id,
+      iscastrated: animal.iscastrated,
+      lifestyleisindoors: animal.lifestyleisindoors,
+      name: animal.name,
+      sex: animal.sex,
+      timeofdeath: animal.timeofdeath,
+      weightingram: animal.weightingram,
+      animalgroupid: animal.fk_animalgroupid,
+      animaltypeid: animal.fk_animaltypeid
+    }));
   },
 
-  async getByName(name: string): Promise<animals> {
+  async getByName(name: string): Promise<AnimalsType> {
     const foundAnimal = await prisma.animals.findFirst({ where: { name } });
 
-    if (!foundAnimal) throw new Error(`Animal not found with name: ${name}`);
+    if (!foundAnimal) {
+      throw new ResourceNotFoundError(`Animal not found with name: ${name}`, 'name', name);
+    }
 
-    return foundAnimal;
+    return ({
+      dateofbirth: foundAnimal.dateofbirth,
+      dateofbirthisexact: foundAnimal.dateofbirthisexact,
+      heightincm: foundAnimal.heightincm,
+      id: foundAnimal.id,
+      iscastrated: foundAnimal.iscastrated,
+      lifestyleisindoors: foundAnimal.lifestyleisindoors,
+      name: foundAnimal.name,
+      sex: foundAnimal.sex,
+      timeofdeath: foundAnimal.timeofdeath,
+      weightingram: foundAnimal.weightingram,
+      animalgroupid: foundAnimal.fk_animalgroupid,
+      animaltypeid: foundAnimal.fk_animaltypeid
+    });
   },
 
-  async getAll(): Promise<animals[]> {
-    return await prisma.animals.findMany();
+  async getAll(): Promise<AnimalsType[]> {
+    const foundAnimals = await prisma.animals.findMany();
+    return foundAnimals.map(foundAnimal => ({
+      dateofbirth: foundAnimal.dateofbirth,
+      dateofbirthisexact: foundAnimal.dateofbirthisexact,
+      heightincm: foundAnimal.heightincm,
+      id: foundAnimal.id,
+      iscastrated: foundAnimal.iscastrated,
+      lifestyleisindoors: foundAnimal.lifestyleisindoors,
+      name: foundAnimal.name,
+      sex: foundAnimal.sex,
+      timeofdeath: foundAnimal.timeofdeath,
+      weightingram: foundAnimal.weightingram,
+      animalgroupid: foundAnimal.fk_animalgroupid,
+      animaltypeid: foundAnimal.fk_animaltypeid
+    }));
   },
 
   async delete(id: number): Promise<void> {
+    const openAppointments = await prisma.appointments.findMany({
+      where: {
+        fk_animalid: id,
+        starttime: {
+          lt: new Date()
+        }
+      },
+      select: {
+        id: true
+      }
+    });
+    
+    if (openAppointments.length > 0) {
+      throw new ConstraintError("Animal has open appointments", openAppointments.map(x => ({
+        path: x.id.toString(),
+        value: x.id
+      })));
+    }
+
     await prisma.animals.delete({ where: { id } });
   },
 
@@ -108,6 +204,7 @@ export const animalService = {
     const filepath = found?.picturepath ?? 'public/placeholders/animal.png';
     return path.join(appRootDir, filepath);
   },
+
   async savePicture(animalId: number, fileOnDiskPath: string | null): Promise<void> {
     const old = await prisma.animals.findFirst({
       where: {
@@ -118,7 +215,7 @@ export const animalService = {
       }
     });
     if (!old) {
-      throw new Error(`Animal with ID(${animalId}) does not exist.`);
+      throw new ConstraintError(`No animal with given id exists.`, [{ path: 'animalId', value: animalId }]);
     }
 
     if (old.picturepath) {
@@ -139,5 +236,19 @@ export const animalService = {
         picturepath: true
       },
     });
+  },
+
+  async canPersonAccessAnimal(personId: number, animalId: number): Promise<boolean> {
+    const relationExists = await prisma.person_has_animal.findFirst({
+      where: {
+        fk_animalid: animalId,
+        fk_personid: personId
+      }
+    });
+    if (relationExists) {
+      return true;
+    }
+
+    return false;
   }
 };
